Add Tag Helpers to a few links

+++ StocksAppAssignment/Models/StockTrade.cs
@@ -9,6 +9,6 @@
 
         public double Price { get; set; }
 
-        public uint? Quantity { get; set; }
+        public double? Quantity { get; set; }
     }
 }
